projectly_backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.2
django-filter==25.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.0
drf-dynamic-fields==0.4.0
drf-excel==2.5.3
drf-extensions==0.8.0
drf-flex-fields==1.0.2
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
Markdown==3.8
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pycparser==2.22
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
gunicorn==20.1.0
whitenoise==6.5.0
django-allauth==0.57.0
django-environ==0.11.2
django-debug-toolbar==4.2.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-tables2==2.7.0
markdown==3.8